requirements.txt
python-telegram-bot>=21.0
aiohttp>=3.8.1
python-dotenv>=1.0.0
loguru>=0.7.0
